# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/plans:
    get:
      operationId: listAdminPlans
      tags: [admin]
      summary: List plans (admin)
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createAdminPlan
      tags: [admin]
      summary: Create plan (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    patch:
      operationId: updateAdminPlan
      tags: [admin]
      summary: Update plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deleteAdminPlan
      tags: [admin]
      summary: Delete plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/platform-configs:
    get:
      operationId: listPlatformConfigs
      tags: [admin]
      summary: List social platform configs
      responses:
        "200":
          description: Platform configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformConfig"
    post:
      operationId: upsertPlatformConfig
      tags: [admin]
      summary: Create or update platform config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformConfigInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformConfig"
  /admin/platform-configs/{platform}:
    get:
      operationId: getPlatformConfig
      tags: [admin]
      summary: Get platform config
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Platform config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformConfig"
components:
  schemas:
...
    Plan:
      type: object
      required: [id, name, priceMonthly, features, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        priceMonthly:
          type: number
        priceYearly:
          type: ["number", "null"]
        currency:
          type: string
        maxSocialAccounts:
          type: ["integer", "null"]
        maxScheduledPosts:
          type: ["integer", "null"]
        maxAiImages:
          type: ["integer", "null"]
        maxTeamMembers:
          type: ["integer", "null"]
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        description:
          type: ["string", "null"]
    PlanInput:
      type: object
      required: [name, priceMonthly, features]
      properties:
        name:
          type: string
        slug:
          type: string
        priceMonthly:
          type: number
        priceYearly:
          type: number
        currency:
          type: string
        maxSocialAccounts:
          type: integer
        maxScheduledPosts:
          type: integer
        maxAiImages:
          type: integer
        maxTeamMembers:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        description:
          type: string
    PlanUpdate:
      type: object
      properties:
        name:
          type: string
        priceMonthly:
          type: number
        priceYearly:
          type: number
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        description:
          type: string
        maxSocialAccounts:
          type: integer
        maxScheduledPosts:
          type: integer

    # ─── SUBSCRIPTION SCHEMAS ──────────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, email, name, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
        isVerified:
          type: boolean
        planName:
          type: ["string", "null"]
        workspacesCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
    AdminUserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended]
        role:
          type: string
          enum: [user, admin]
        planId:
          type: integer
    PlatformConfig:
      type: object
      required: [id, platform, isEnabled, environment]
      properties:
        id:
          type: integer
        platform:
          type: string
        clientId:
          type: ["string", "null"]
        redirectUri:
          type: ["string", "null"]
        scopes:
          type: ["string", "null"]
        environment:
          type: string
          enum: [production, sandbox]
        isEnabled:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    PlatformConfigInput:
      type: object
      required: [platform]
      properties:
        platform:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        redirectUri:
          type: string
        scopes:
          type: string
        webhookSecret:
          type: string
        environment:
          type: string
          enum: [production, sandbox]
        isEnabled:
          type: boolean