# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SocialFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  slug?: string;
  priceMonthly: number;
  /** @nullable */
  priceYearly?: number | null;
  currency?: string;
  /** @nullable */
  maxSocialAccounts?: number | null;
  /** @nullable */
  maxScheduledPosts?: number | null;
  /** @nullable */
  maxAiImages?: number | null;
  /** @nullable */
  maxTeamMembers?: number | null;
  features: string[];
  isActive: boolean;
  isFeatured?: boolean;
  /** @nullable */
  description?: string | null;
}