# lib/api-zod/src/generated/types/planUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SocialFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlanUpdate {
  name?: string;
  priceMonthly?: number;
  priceYearly?: number;
  features?: string[];
  isActive?: boolean;
  isFeatured?: boolean;
  description?: string;
  maxSocialAccounts?: number;
  maxScheduledPosts?: number;
}